Add index driver abstraction layer.
